+++ src/polymesh/pm.hh
@@ -1,11 +1,11 @@
 #pragma once
 
-// Includes all polymesh features
+// Includes most commonly used polymesh features
 
 #include <polymesh/Mesh.hh>
 
-#include <polymesh/algorithms/optimization.hh>
-#include <polymesh/algorithms/properties.hh>
+#include <polymesh/properties.hh>
 
 #include <polymesh/formats.hh>
+
 #include <polymesh/objects.hh>